requirements/base.txt
Django==5.2.5

dj-database-url==3.0.1
python-decouple==3.8

crispy-bootstrap5==2025.6
django-crispy-forms==2.4

whitenoise==6.9.0
